v2.1 updates - See notes
QUESTION FONT: Please make the font larger for all form questions (22px). I think it's easy to skip over a question because the font is kind of small/similar to the notes on the rest of the form.

This has been done, and was a great recommendation!

STRAP COLOR REFERENCE PHOTO - can you make it clickable to zoom/enlarge? Or maybe have a link that says "Click here for color reference" that would bring up this photo.

This has been implemented, let me know if you see anything that should be changed!

STRAP COLOR: Update the color "Tan (Jute)" to "Tan (Old Gold)"

This is done

SPORT OR WORD QUESTION: Change the example text in the "Sport or Word on Strap" field to say "Example: "VOLLEYBALL" or "PRACTICE" or "TRAVEL" and then change the note below the field to say "Note: For best results, limit to 20 characters"

This is done. I actually limited the text entry limit for this question and the "Names" fields to be no more than 20 characters. The customer will see a character count, and messaging about this max limit.

NUMBERS: Is there any way to make this section work so that when choosing numbers, you can hold Shift and select multiples? Not a major concern, but it would be nice!

Yes! This was a great suggestion and has been implemented.

FONT PICKER: I'm glad to hear that PlayerKits is on board with the customer adding their own fonts! I think there will be some schools that have already uploaded their font, and they just want us to use it - can you add an option for the customer to indicate "Use Font Previously Provided" and "I Need to Upload a New Font" (which would then allow the file upload) or something like that...

I have included a new dropdown for this question with the options "Use Default Font(s)", "Use Font(s) Previously Provided", and "I Need to Upload a New Font". The custom font upload field is only displayed if the customer selects "I Need to Upload a New Font" from the dropdown. Also improved the field hints/messages.

FONT COLOR: In the font color picker section, is it better to just have a field where the customer can type in the color? This is really cool functionality, but I think it might be overly complicated. Whatever the color is, it is likely going to be one that references the color in the logo - which PlayerKits can sample. I worry that having Hex codes and RGB will be too overwhelming for many of our customers. What do you think?

This is a tough one. On one hand, I would think that they would know the proper CMYK, RGB or HEX value for their school or team's colors. But on the other hand, I see your point, and some people ordering may be less informed or know these values. Since I've already coded this, let's give it a shot as-is for awhile and see how customers interact with it.

On a side note, did you notice that if you upload a logo file, it extracts the colors from the image and presents them as options for the font color? Do you like this part of the functionality?

NUMBER OF SETS: Jen thinks we should add more options to the "Number of Sets" question. Maybe make this a dropdown like you have the "Blanks" question.

I have changed this to be a dropdown, with the numbers 1-25 as options. Let me know if I need to add more options here.

LOOP PREVIEW: Jen was concerned about this (Wayne absolutely LOVES this functionality, so it's definitely staying!) but I'm fine to keep it. If customers are confused by it, we might consider taking it down at a later date. That said, I think we need to:
Make the disclaimer text "Preview is for visualization purposes only. Actual product may vary slightly" larger and RED.

I made this 16px and red, great suggestion.

It might have been too crazy to do, but I thought you said you could change the preview to show the shorter tag image for Numbers and the larger tag image for Names. Not a huge deal, just asking!

I have now implemented this functionality as follows:

If the customer selects "Names", a large loop preview image is shown (regardless of whether they selected a "Single" or "Double" clip above).

If the customer selects "Numbers", a small loop preview image is shown for "Single" clip loops, and a large loop preview image is shown for "Double" clip loops.

Let me know if you need/want any changes to this logic/functionality

ADD TO CART / CHECKOUT / NEXT STEPS: This section keeps confusing me and I'm getting caught up here. I think it's because I already pressed "Add to Cart" so it kind of seems like it didn't work. Let's try this:
After adding an order to the cart, can you:
Hide the "Clear Form" and "Add to Cart" buttons above the success message, as well as:

Great suggestion, these are now hidden after clicking the "Add to Cart" button.

Include another green button to the left that says "Start a New Loop Order" - this button would essentially clear the form and take them to the top.  What do you think?

Also a great suggestion, this has been implemented as well as a "View Cart" button that can take them to their cart. I also added success messaging that says "Your custom loops have been added to the cart!".

MINIMUM ORDER FEE: Jen mentioned a "Minimum Order Fee" for orders under 30. Does this form/process account for that?

I implemented this the way that Wayne asked me to. He asked me to do the following:

If the customer orders less than 24 loops, the per-product sublimation fee is $3.95/ea

If the customer orders 24 or more loops, the per-product sublimation fee is $1.95/ea

If this is different than what Jen wants or is expecting, please ask her to consult Wayne about this. If they decide together to make a change, I'll update the code to match.